change batch_size,seq_len=2

+++ train.py
@@ -18,7 +18,7 @@ parser.add_argument('--gpu_ids', type=str, default='1', help='gpu ids: e.g. 0  0
 parser.add_argument('--save_dir', type=str, default='./results', help='path to save the result')
 
 parser.add_argument('--train_seq', type=list, default=['00', '01', '02', '04', '06', '08', '09'], help='sequences for training')
-parser.add_argument('--val_seq', type=list, default=['05', '07', '10'], help='sequences for validation')
+parser.add_argument('--val_seq', type=list, default=['05', '07', '10', '01', '02', '04'], help='sequences for validation')
 parser.add_argument('--seed', type=int, default=0, help='random seed')
 
 parser.add_argument('--img_w', type=int, default=512, help='image width')
@@ -128,7 +128,7 @@ def train(model, optimizer, train_loader, selection, temp, logger, ep, p=0.5, we
 
         optimizer.zero_grad()
         # 模型结果
-        poses, attn_map, decisions, probs = model(imgs)
+        poses, decisions, probs = model(imgs)
 
         
         if not weighted:
@@ -270,7 +270,7 @@ def main():
         avg_pose_loss, avg_penalty_loss = train(model, optimizer, train_loader, selection, temp, logger, ep, p=0.5)
         
         # Save the model after training
-        if(os.path.isfile(f'{checkpoints_dir}/{(ep-1):003}.pth')):
+        if(os.path.isfile(f'{checkpoints_dir}/{(ep-1):003}.pth')) and ep%10!=0:
             os.remove(f'{checkpoints_dir}/{(ep-1):003}.pth')
         torch.save(model.module.state_dict(), f'{checkpoints_dir}/{ep:003}.pth')
         # Save the model after training
@@ -278,7 +278,7 @@ def main():
         print(message)
         logger.info(message)
         
-        if ep > args.epochs_warmup+args.epochs_joint or (ep > 10 and ep % 2 == 0):
+        if ep > args.epochs_warmup+args.epochs_joint or ep%10==0:
             # Evaluate the model
             print('Evaluating the model')
             logger.info('Evaluating the model')
@@ -314,6 +314,25 @@ def main():
             print(message)
             if args.experiment_name != 'debug':
                 wandb.log({"10. t_rel": round(errors[2]['t_rel'], 4), "10. r_rel": round(errors[2]['r_rel'], 4), "10. t_rmse": round(errors[2]['t_rmse'], 4), "10. r_rmse": round(errors[2]['r_rmse'], 4)})
+            
+            message = "Epoch {} evaluation Seq. 01 , t_rel: {}, r_rel: {}, t_rmse: {}, r_rmse: {}" .format(ep, round(errors[2]['t_rel'], 4), round(errors[2]['r_rel'], 4), round(errors[2]['t_rmse'], 4), round(errors[2]['r_rmse'], 4))
+            logger.info(message)
+            print(message)
+            if args.experiment_name != 'debug':
+                wandb.log({"01. t_rel": round(errors[2]['t_rel'], 4), "01. r_rel": round(errors[2]['r_rel'], 4), "01. t_rmse": round(errors[2]['t_rmse'], 4), "01. r_rmse": round(errors[2]['r_rmse'], 4)})
+            
+            message = "Epoch {} evaluation Seq. 02 , t_rel: {}, r_rel: {}, t_rmse: {}, r_rmse: {}" .format(ep, round(errors[2]['t_rel'], 4), round(errors[2]['r_rel'], 4), round(errors[2]['t_rmse'], 4), round(errors[2]['r_rmse'], 4))
+            logger.info(message)
+            print(message)
+            if args.experiment_name != 'debug':
+                wandb.log({"02. t_rel": round(errors[2]['t_rel'], 4), "02. r_rel": round(errors[2]['r_rel'], 4), "02. t_rmse": round(errors[2]['t_rmse'], 4), "02. r_rmse": round(errors[2]['r_rmse'], 4)})
+            
+            message = "Epoch {} evaluation Seq. 04 , t_rel: {}, r_rel: {}, t_rmse: {}, r_rmse: {}" .format(ep, round(errors[2]['t_rel'], 4), round(errors[2]['r_rel'], 4), round(errors[2]['t_rmse'], 4), round(errors[2]['r_rmse'], 4))
+            logger.info(message)
+            print(message)
+            if args.experiment_name != 'debug':
+                wandb.log({"04. t_rel": round(errors[2]['t_rel'], 4), "04. r_rel": round(errors[2]['r_rel'], 4), "04. t_rmse": round(errors[2]['t_rmse'], 4), "04. r_rmse": round(errors[2]['r_rmse'], 4)})
+            
             logger.info(message)
             print(message)
